requirements.txt
alembic==1.12.0
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.3
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.0.5
flask-restx==1.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.2
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
referencing==0.30.2
rpds-py==0.10.4
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==3.0.0
